+++ lib/Phpactor.php
@@ -14,11 +14,13 @@ use Phpactor\Extension\Navigation\NavigationExtension;
 use Phpactor\Extension\SourceCodeFilesystemExtra\SourceCodeFilesystemExtraExtension;
 use Phpactor\Extension\SourceCodeFilesystem\SourceCodeFilesystemExtension;
 use Phpactor\Extension\WorseReflectionExtra\WorseReflectionExtraExtension;
+use Phpactor\Extension\ExtensionManager\ExtensionManagerExtension;
 use Phpactor\Extension\WorseReflection\WorseReflectionExtension;
 use Phpactor\Extension\ClassMover\ClassMoverExtension;
 use Phpactor\FilePathResolverExtension\FilePathResolverExtension;
 use Phpactor\Extension\ContextMenu\ContextMenuExtension;
 use Phpactor\Extension\Rpc\RpcExtension;
+use Phpactor\Extension\LanguageServer\LanguageServerExtension;
 use Phpactor\Extension\Console\ConsoleExtension;
 use Phpactor\Extension\ClassToFile\ClassToFileExtension;
 use Phpactor\Extension\Logger\LoggingExtension;
@@ -30,7 +32,6 @@ use Symfony\Component\Console\Input\InputInterface;
 use Phpactor\Config\Paths;
 use Phpactor\Extension\ClassToFileExtra\ClassToFileExtraExtension;
 use Composer\XdebugHandler\XdebugHandler;
-use Phpactor\Extension\LanguageServer\LanguageServerExtension;
 
 class Phpactor
 {
@@ -42,6 +43,10 @@ class Phpactor
         $config = $configLoader->loadConfig();
         $config[CoreExtension::COMMAND] = $input->getFirstArgument();
         $config[FilePathResolverExtension::PARAM_APPLICATION_ROOT] = realpath(__DIR__ . '/..');
+        $config[ExtensionManagerExtension::PARAM_EXTENSION_VENDOR_DIR] = $extensionDir = __DIR__ . '/../vendor/phpactor/extensions';
+        $config[ExtensionManagerExtension::PARAM_INSTALLED_EXTENSIONS_FILE] = $extensionDir. '/extensions.php';
+        $config[ExtensionManagerExtension::PARAM_VENDOR_DIR] = realpath(__DIR__ . '/../vendor');
+        $config[ExtensionManagerExtension::PARAM_EXTENSION_CONFIG_FILE] = $extensionDir .'/extensions.json';
 
         $cwd = getcwd();
 
@@ -76,7 +81,8 @@ class Phpactor
             LoggingExtension::class,
             ComposerAutoloaderExtension::class,
             ConsoleExtension::class,
-            LanguageServerExtension::class
+            LanguageServerExtension::class,
+            ExtensionManagerExtension::class,
         ];
 
         $container = new PhpactorContainer();