+++ CHANGELOG.md
@@ -15,6 +15,7 @@ BC Break:
 
 Features:
 
+  - [ExtensionManager] Facility to dynamically add extensions to Phpactor
   - [Completion] Class alias completor, #592
   - [CodeTransform] Cycle class member visiblity #521
   - [LSP] Experimental language server support (completion and goto